Split tests into separate files per code host
